Code Refactor

+++ energy_analytics_service/src/main/java/com/fhdo/energy_analytics_service/domain/entity/EnergyTypeEntity.java
@@ -18,5 +18,5 @@ public class EnergyTypeEntity {
 	private String provider;
 	private String unit;
 	@Column(name = "price_per_unit")
-	private Integer pricePerUnit;
+	private double pricePerUnit;
 }
